+++ bdc/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Now implemented in Python 3. Will _not_ work with Python 2.
 - Added support for arbitrary build profiles.
 - Added support for instructor note consolidation
+- Added support for an include directive in the `build.yaml` file 
+  (e.g., `#include "file.yaml"`)
 
 ### Version 1.29.0